refactor: move services to app layer
