+++ examples/CMakeLists.txt
@@ -8,8 +8,8 @@ add_executable(rabbitmq_example RabbitMQExample.cpp ../src/storage/mq/RabbitMQCl
 target_include_directories(rabbitmq_example PUBLIC ${CMAKE_SOURCE_DIR}/src/storage/mq ${CMAKE_SOURCE_DIR}/src/framework/utils)
 target_link_libraries(rabbitmq_example muduo_core yaml-cpp ${LIBS})
 
-add_executable(web_server WebServer.cpp ../src/services/UserService.cpp ../src/storage/UserRepository.cpp ../src/framework/ioc/Container.cpp)
-target_include_directories(web_server PRIVATE ${CMAKE_SOURCE_DIR}/src/framework/ioc ${CMAKE_SOURCE_DIR}/src/services ${CMAKE_SOURCE_DIR}/src/storage)
+add_executable(web_server WebServer.cpp ../src/app/UserService.cpp ../src/storage/UserRepository.cpp ../src/framework/ioc/Container.cpp)
+target_include_directories(web_server PRIVATE ${CMAKE_SOURCE_DIR}/src/framework/ioc ${CMAKE_SOURCE_DIR}/src/app ${CMAKE_SOURCE_DIR}/src/storage)
 target_link_libraries(web_server muduo_core ${LIBS})
 
 add_executable(redis_example RedisExample.cpp ../src/storage/cache/RedisClient.cpp ../src/storage/cache/RedisPool.cpp ../src/framework/utils/ConfigManager.cpp)
@@ -29,7 +29,7 @@ target_include_directories(session_example PRIVATE ${CMAKE_SOURCE_DIR}/src/frame
 target_link_libraries(session_example muduo_http)
 
 add_executable(web_server WebServer.cpp)
-target_include_directories(web_server PRIVATE ${CMAKE_SOURCE_DIR}/src/core/include ${CMAKE_SOURCE_DIR}/src/modules ${CMAKE_SOURCE_DIR}/src/framework ${CMAKE_SOURCE_DIR}/src/services)
+target_include_directories(web_server PRIVATE ${CMAKE_SOURCE_DIR}/src/core/include ${CMAKE_SOURCE_DIR}/src/modules ${CMAKE_SOURCE_DIR}/src/framework ${CMAKE_SOURCE_DIR}/src/app)
 target_link_libraries(web_server muduo_http muduo_core ${LIBS})
 
 add_executable(logger_stress LoggerStress.cpp ../src/framework/utils/ConfigManager.cpp)
